refactor ILocation interface into class
I need to add methods to the location objections being used within the
LocationDatabase. The best course was to make them into Location class
instances that implement the ILocation interface.

Also complying with "one class per file" per tslint, so locations is now
a directory with Location/ILocation in one file and LocationDatabase
in its own.