refactor public LocationDatabase API to use ILocation
Partial reversion of f15d4fd

It was a mistake to change the public API to use Location instances when
I refactored in f15d4fd. It's a better practice to use the interface
anyway, and it's proven problematic as I want to add private members to
Location objects.

I'm also re-consolidating ILocation and Location back into the same file
as LocationDatabase. I'm defying tslint here by having two classes in
one file, but without package privacy, I couldn't find a way to add
private data to the Location class without having to expose it publicly
unless it was in the same file.